split DiscriminatedType into the right package
